Fix search not working properly
Previously, we would match a student to an input name if only one
part of the input name matched with one part of the student name.
This means that we would often match with too many students so the
app wouldn't work for those who have a common name.

Now, we compute a "match score" for every student name and we return
the student with the highest score, except when there are two equal
scores in which case the user needs to type more to disambiguate the
student that they are looking for.