Revert caching class data
Firstly, there was a bug where the newly fetched data was only shown
after the page was reloaded, and this bug would be hard to fix without
rerendering the whole interface twice for each load, which I don't
like because the site is supposed to be used on mobile devices.

Secondly, it can be confusing because if the fetch of the data takes
too much time, then the user doesn't know that they might be looking
at stale data.