Add background_color to webmanifest
The background_color is the color that is shown behind the application
logo during app launch, before the page has finished loading.

I think this property has to be set for Chrome to show the PWA
install prompt on the website: previously, it was not showing, and now
I hope it will be showing.

After the page is loaded, the actual page background color will be
either #ffffff or #000000 depending on the prefered color scheme,
so in case the user has chosen a light color scheme, there will be an
ugly transition between #000000 and #ffffff. I chose to do it this way
and not the other way around (a #ffffff background color that would be
ugly for those who have chosen a dark color scheme) because I believe
that it is better for the eyes to switch from launch screen to black to
white than it is to switch from launch screen to white to black,
especially in low light environments.