Prerender viewer HTML with last search result
When the user types a new query, after the search results are shown we
send the document's HTML to the Service Worker so that it serves it the
next time the page is loaded. This means that even before the JavaScript
is loaded on the next refresh, the user will already be able to see the
last search result, although the relative times might be off.