Automatically change the Service Worker when needed
We now generate a hash of the resources that the SW caches and include
it inside of the SW.

This will prevent issues where someone forgets to update the version as
there is no longer a version to update. This also makes it easier for
other contributors to edit files.